# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  routes.FTCMatch:
    properties:
      actualStartTime:
        type: string
      description:
        type: string
      matchNumber:
        type: integer
      startTime:
        type: string
      teams:
        items:
          properties:
            station:
              type: string
            teamNumber:
              type: integer
          type: object
        type: array
      tournamentLevel:
        type: string
    type: object
  routes.FTCTeam:
    properties:
      city:
        type: string
      country:
        type: string
      nameFull:
        type: string
      nameShort:
        type: string
      stateprov:
        type: string
      teamNumber:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Blue Banner Engine, providing FRC match
    predictions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blue Banner Engine (BBE) API
  version: "1.0"
paths:
...
  /ftc/event/{season}/{eventCode}/matches:
    get:
      description: Retrieves the match schedule for an FTC event from the FIRST FTC
        API
      parameters:
      - description: FTC Season (e.g., 2024)
        in: path
        name: season
        required: true
        type: string
      - description: FTC Event Code (e.g., TXHOU)
        in: path
        name: eventCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.FTCMatch'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get FTC Event Match Schedule
      tags:
      - ftc
  /ftc/event/{season}/{eventCode}/team/{teamNumber}/matches:
    get:
      description: Retrieves matches for a specific team at an FTC event
      parameters:
      - description: FTC Season (e.g., 2024)
        in: path
        name: season
        required: true
        type: string
      - description: FTC Event Code (e.g., TXHOU)
        in: path
        name: eventCode
        required: true
        type: string
      - description: Team Number
        in: path
        name: teamNumber
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.FTCMatch'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get FTC Team Matches
      tags:
      - ftc
  /ftc/event/{season}/{eventCode}/teams:
    get:
      description: Retrieves the list of teams at an FTC event from the FIRST FTC
        API
      parameters:
      - description: FTC Season (e.g., 2024)
        in: path
        name: season
        required: true
        type: string
      - description: FTC Event Code (e.g., TXHOU)
        in: path
        name: eventCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.FTCTeam'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get FTC Event Teams
      tags:
      - ftc
...
swagger: "2.0"