# main.go
// main.go

package main

import (
	"log"
	"net/http"
	"os"
	"strings"
	"time"

	"blue-banner-engine/docs"
	pb "blue-banner-engine/protos"
	pb_analytics "blue-banner-engine/protos/analytics"

	"path/filepath"

	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	"github.com/joho/godotenv"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	"google.golang.org/grpc"
	"google.golang.org/grpc/credentials/insecure"

	"blue-banner-engine/src/routes"
)

type ErrorResponse struct {
	Error string `json:"error"`
}

type ShapAnalysisJSON struct {
	BaseValue    float32   `json:"base_value"`
	Values       []float32 `json:"values"`
	FeatureNames []string  `json:"feature_names"`
	FeatureData  []float32 `json:"feature_data"`
}

type PredictionResponseJSON struct {
	MatchKey        string             `json:"match_key" example:"2025mxle_qm1"`
	PredictedWinner string             `json:"predicted_winner" example:"blue"`
	WinProbability  map[string]float32 `json:"win_probability" example:"red:0.21,blue:0.79"`
	PredictedScores map[string]int32   `json:"predicted_scores" example:"red:95,blue:112"`
	Status          string             `json:"status" example:"played"`
	ActualWinner    string             `json:"actual_winner,omitempty" example:"red"`
	ActualScores    map[string]int32   `json:"actual_scores,omitempty" example:"red:100,blue:98"`
	ShapAnalysis    *ShapAnalysisJSON  `json:"shap_analysis,omitempty"`
}

type PredictionHandler struct {
	grpcClient pb.MatchpointClient
	tbaApiKey  string // Almacenamos la API key
}

type TbaEvent struct {
	Key       string `json:"key"`
	Name      string `json:"name"`
	EventCode string `json:"event_code"`
	EventType int    `json:"event_type"`
	City      string `json:"city"`
	StateProv string `json:"state_prov"`
	Country   string `json:"country"`
	StartDate string `json:"start_date"`
	EndDate   string `json:"end_date"`
}

// @title           Blue Banner Engine (BBE) API
// @version         1.0
// @description     This is the API server for the Blue Banner Engine, providing FRC match predictions.
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.url    http://www.swagger.io/support
// @contact.email  support@swagger.io

// @license.name  MIT
// @license.url   https://opensource.org/licenses/MIT

// @BasePath  /api/v1
func main() {

	err := godotenv.Load() // Load .env file from current directory
	if err != nil {
		log.Println("Error loading .env file")
	}
	tbaKey := os.Getenv("TBA_API_KEY")
	if tbaKey == "" {
		log.Println("WARNING: TBA_API_KEY environment variable not set. Real match data will not be available.")
	}

	ftcKey := os.Getenv("FTC_API_KEY")
	ftcUser := os.Getenv("FTC_USER")
	if ftcKey == "" || ftcUser == "" {
		log.Println("WARNING: FTC_API_KEY or FTC_USER environment variable not set. FTC event data will not be available.")
	}

	grpcServerAddress := os.Getenv("GRPC_SERVER_ADDRESS")
	if grpcServerAddress == "" {
		grpcServerAddress = "localhost:50051"
	}

	// gRPC Client Configuration
	conn, err := grpc.Dial(grpcServerAddress, grpc.WithTransportCredentials(insecure.NewCredentials()))
	if err != nil {
		log.Fatalf("did not connect: %v", err)
	}
	defer conn.Close()

	handler := &PredictionHandler{
		grpcClient: pb.NewMatchpointClient(conn),
		tbaApiKey:  tbaKey,
	}
	analyticsClient := pb_analytics.NewAnalyticsClient(conn)

	router := gin.Default()

	router.Use(cors.New(cors.Config{
		AllowOrigins:     []string{"http://localhost:5173", "http://localhost:8080", "https://bbe-frc.com"},
		AllowMethods:     []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowHeaders:     []string{"Origin", "Content-Type", "Accept", "Authorization"},
		ExposeHeaders:    []string{"Content-Length"},
		AllowCredentials: true,
		MaxAge:           12 * time.Hour,
	}))

	// Configure Swagger host based on environment variable
	swaggerHost := os.Getenv("SWAGGER_HOST")
	if swaggerHost == "" {
		swaggerHost = "localhost:8080" // default
	}
	docs.SwaggerInfo.Host = swaggerHost
	docs.SwaggerInfo.BasePath = "/api/v1"

	v1 := router.Group("/api/v1")
	{
		// ML and Analytics routes (no auth required for now)
		// ML and Analytics routes (no auth required for now)
		routes.RegisterMatchpointRoutes(v1, handler.grpcClient, tbaKey)
		routes.RegisterTbaRoutes(v1, tbaKey)
		routes.RegisterFtcRoutes(v1, ftcUser, ftcKey)
		routes.RegisterSimulationRoutes(v1, handler.grpcClient)
		routes.RegisterAnalyticsRoutes(v1, analyticsClient)

		// New Metrics Routes
		routes.RegisterStatboticsRoutes(v1)
		routes.RegisterMetricRoutes(v1)
	}

	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// Serve email templates
	// router.Static("/static/templates", "./static/templates")

	staticFiles := "bbe-ui/dist"
	templatesFiles := "supabase/email-templates"

	router.StaticFS("/static", http.Dir(filepath.Join(staticFiles, "static")))
	router.StaticFS("/templates", http.Dir(filepath.Join(templatesFiles)))
	router.StaticFS("/assets", http.Dir(filepath.Join(staticFiles, "assets")))
	router.StaticFile("/manifest.json", filepath.Join(staticFiles, "manifest.json"))
	router.StaticFile("/bbe_logo_no_name.png", filepath.Join(staticFiles, "bbe_logo_no_name.png"))
	router.StaticFile("/favicon.ico", filepath.Join(staticFiles, "bbe_logo_no_name.png"))

	router.GET("/favicon.png", func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept, Authorization")
		c.File("docs/visuals/bbe_logo_no_name.png") // Adjust path if needed
	})

	router.GET("/swagger.json", func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept, Authorization")
		c.File("docs/swagger.json")
	})
	router.OPTIONS("/swagger.json", func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept, Authorization")
		c.Status(http.StatusOK)
	})

	router.NoRoute(func(c *gin.Context) {
		if !strings.HasPrefix(c.Request.URL.Path, "/api") && !strings.HasPrefix(c.Request.URL.Path, "/swagger") && !strings.HasPrefix(c.Request.URL.Path, "/favicon.ico") {
			path := filepath.Join(staticFiles, "index.html")
			_, err := os.Stat(path)
			if os.IsNotExist(err) {
				c.Writer.WriteHeader(http.StatusNotFound)
				return
			}
			c.File(path)
		}
	})

	log.Println("Starting BBE server on port 8080...")
	log.Println("Access the UI at https://bbe-frc.com")
	log.Println("API documentation at https://bbe-frc.com/swagger/index.html")

	if os.Getenv("PRODUCTION") == "true" &&
		os.Getenv("CERT_PATH") != "" &&
		os.Getenv("KEY_PATH") != "" {
		log.Println("Running on production")
		router.RunTLS(":8080", os.Getenv("CERT_PATH"), os.Getenv("KEY_PATH"))
	} else {
		log.Println("Running on testing")

		for _, item := range router.Routes() {
			println("method:", item.Method, "path:", item.Path)
		}
		router.Run(":8080")
	}
}